Refactoring how the Inspector works with commands.
Filters are run after type validation.
Adding a functional test to ensure some of the JSON example service
description use cases will work.
Instead of the slew of boolean options the validateConfig() method
required, I added a initConfig() method to add default and static values
to a collection, and removed all validation and idempotent specific args
from validateConfig().  When a commmand is instantiated, default and
static values are added as appropriate.  When a command is being
prepared, defaults and statics are again added, but parameters are also
validated (assuming type validation was not globally disabled on the
inspector).