Fix RetryMiddleware default exponential delay (#2132)
* Fix RetryMiddleware::exponentialDelay by making it return milliseconds

RetryMiddleware::exponentialDelay was being used as-is to set the delay option in the retry middleware. However, the option requires milliseconds and this method was returning seconds.

* Fix RetryMiddleware::exponentialDelay by making it return milliseconds

RetryMiddleware::exponentialDelay was being used as-is to set the delay option in the retry middleware. However, the option requires milliseconds and this method was returning seconds.

* Fix exponentialDelay int casting

* Added comment